backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Auth & security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Stripe
stripe>=8.0.0

# Storage (Supabase)
supabase>=2.3.0

# Email (SendGrid)
sendgrid>=6.11.0

# Monitoring
sentry-sdk[fastapi]>=1.39.0

# Utils
python-dotenv>=1.0.0
